# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server.",
        "title": "My API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/",
    "paths": {
        "/result/{task_id}": {
            "get": {
                "description": "Post task_id in DataBase",
                "summary": "Get task result",
                "parameters": [
                    {
                        "type": "string",
                        "description": "task_id",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "result\": \"result state",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "404": {
                        "description": "Failed to get result",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/status/{task_id}": {
            "get": {
                "description": "Post task_id in DataBase",
                "summary": "Get task status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "task_id",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status\": \"status state",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "404": {
                        "description": "Failed to get status",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/task": {
            "post": {
                "description": "Post task_id in DataBase and start task",
                "summary": "Post task task_id",
                "responses": {
                    "201": {
                        "description": "task_id\": \"id value",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "404": {
                        "description": "Failed to store value",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}